Add test for reading addresses from available adapters
Moving all tests to dfu-test.js for now, so that we can reuse some common
methods. The test that reads addresses from available adapters cause
segfault on Linux. The test that opens and closes adapter fails because
the close operation appears to crash, returning an error code. Skipping
the adapter open/close test and the main DFU test for now, as they will
fail anyway as long as reading of addresses do not work.