Fix occasional OPERATION_NOT_PERMITTED error on resume
Occasionally, we have experienced OPERATION_NOT_PERMITTED errors from
the DFU target when trying to resume a previous DFU transfer.

According to the DFU spec, the target should clear its crc32 value
after EXECUTE. It turns out that this is incorrect. The crc32 value
is only cleared after the final EXECUTE. Our implementation assumed
that a fully written object offset and a non-empty crc32 value always
meant that there was an object on the target that had not been
executed. So our implementation always issued an EXECUTE command in
this scenario.

If there is a fully written object offset on the target, we have no
way of knowing whether EXECUTE has been done or not. We just have to
try executing, and catch the OPERATION_NOT_PERMITTED error if it
occurs.