Return exit code 1 if build fails
The build script should exit with error code 1 if there is something
wrong with the build. If not, the CI server will not detect that the
build has failed, and will just continue as if everything is fine.

The buildSystem.rebuild() function returns a promise. We have to
catch errors from that promise, and return process.exit(1).