Increase max gattc write retries from 10 to 20
When performing DFU with nRF52 kits on 1M baud rate, we have
observed that it occasionally needs more than 10 gattc write
retries. The max number of retries observed during DFU was 12.

We could increase the retry delay, but that would lead to reduced
throughput. Increasing the max number of retries instead, as this
should not affect throughput. Now it will only try a few more times
before bailing out.