Remove unreliable tests in dfuSpeedometer-test.js
dfuSpeedometer.js keeps track of DFU transfer speed and progress.
If startTime is not given in the constructor, and currentTime is
not given in updateState(), it will create Date objects internally and
use these to calculate bytes per second.

If the constructor and updateState() are invoked right after each other,
which may happen in unit tests, the time difference in ms between
startTime and currentTime may be 0. Then it is not possible to calculate
bytes per second, and calculateBytesPerSecond() will return 0. We have
seen failing tests on the CI server because of this.

Unit tests that rely on system time are not reliable, so removing these
tests. It is better to allow the unit tests to inject/mock the time. We
are doing this in other tests in this file, and I believe they provide
enough coverage in this case.