Include error details if requesting ATT MTU fails
The _makeError function in adapter.js accepts a second parameter that,
according to the JSDoc, is used for error details. However, this is not
used consistently. Sometimes a string is given here, and sometimes an
error object is given. This makes it difficult to use, as we don't know
what type we are dealing with.

Including error details in the main message instead.