Align default options for enable BLE
There are two ways of enabling BLE:
- Passing `enableBLE: true` as an option to the `adapter.open()` call
- Calling `adapter.enableBLE()` after opening.

When not specifying options to `adapter.enableBLE()`, an options object
may be specified. If this is not given, then a set of default options
are used. When passing `enableBLE: true` to `adapter.open()` it will
always use default options. However, the default options in this case
are not the same as the default options for `adapter.enableBLE()`.

Now using the same default options for `enableBLE: true` as for
`adapter.enableBLE()`.