Wait for TX complete when doing write w/o response
If a lot of 'write without response' operations are performed in quick
succession, the connectivity device may run out of resources and fail
with NRF_INTERNAL_ERROR. The issue is easily reproducible when doing DFU
from a nRF51 dongle to a nRF52 target with SoftDevice v14.

To avoid this, we now wait until we have received BLE_EVT_TX_COMPLETE
before considering the write operation as done. This makes the write
operation more robust at the cost of reduced throughput.