Add FirmwareRegistry for connectivity firmware info
Allowing consumers of pc-ble-driver-js to query information about the
included connectivity firmwares. This includes information like which
firmware file, baud rate, SoftDevice version to use together with a
given device family.

There is already a FirmwareUpdater class that have some of the same
functionality, but this only supports firmwares for J-Link devices. This
is now deprecated, and will be replaced by the nrf-device-setup library.
Keeping the FirmwareUpdater for backwards compatibility.

The FirmwareRegistry also has a `getDeviceSetup()` function that returns
connectivity firmware information as an object that can be passed
directly to the nrf-device-setup library.