Remove PR #184 due to failed integration tests
Two integration tests in simpleSecurity.test.js fails
using the code in PR #184.

Since we are in a hurry to release a new version of
pc-nrfconnect-core, now we do not have time
to investigate the root cause right now.

We will investigate #184 and apply the changes
after the release of pc-nrfconnect-core.