Don't register vendor specific UUIDs on attribute read
Currently, we register any 128 UUID bases unknown to the soft device during a read operation. This was likely a step in an attempt at optimizing the attribute discovery process, but is ineffective in it self since the discovery process finishes before any read responses are processed. More to the point: It also causes us to overflow the maximum number of registered vendor specific UUIDs when performing discovery on some peripherals, like the SensorBug.

Remove the calls to uuidToDriver from_parseGattcReadResponseEvent.